+++ hc/lib/html.py
@@ -6,13 +6,14 @@ class TextOnlyParser(HTMLParser):
         super().__init__(*args, **kwargs)
         self.active = True
         self.buf = []
+        self.skiplist = set(["script", "style"])
 
     def handle_starttag(self, tag, attrs):
-        if tag in ("script", "style"):
+        if tag in self.skiplist:
             self.active = False
 
     def handle_endtag(self, tag):
-        if tag in ("script", "style"):
+        if tag in self.skiplist:
             self.active = True
 
     def handle_data(self, data):
@@ -24,7 +25,10 @@ class TextOnlyParser(HTMLParser):
         return " ".join(messy.split())
 
 
-def html2text(html):
+def html2text(html, skip_pre=False):
     parser = TextOnlyParser()
+    if skip_pre:
+        parser.skiplist.add("pre")
+
     parser.feed(html)
     return parser.get_text()